+++ fastapi_cache/decorator.py
@@ -1,4 +1,5 @@
 import inspect
+import logging
 import sys
 from functools import wraps
 from typing import Any, Awaitable, Callable, Optional, Type, TypeVar
@@ -15,6 +16,8 @@ from starlette.responses import Response
 from fastapi_cache import FastAPICache
 from fastapi_cache.coder import Coder
 
+logger = logging.getLogger(__name__)
+logger.addHandler(logging.NullHandler())
 P = ParamSpec("P")
 R = TypeVar("R")
 
@@ -125,7 +128,8 @@ def cache(
                 )
             try:
                 ttl, ret = await backend.get_with_ttl(cache_key)
-            except ConnectionError:
+            except Exception:
+                logger.warning(f"Error retrieving cache key '{cache_key}' from backend:", exc_info=True)
                 ttl, ret = 0, None
             if not request:
                 if ret is not None:
@@ -133,8 +137,8 @@ def cache(
                 ret = await ensure_async_func(*args, **kwargs)
                 try:
                     await backend.set(cache_key, coder.encode(ret), expire)
-                except ConnectionError:
-                    pass
+                except Exception:
+                    logger.warning(f"Error setting cache key '{cache_key}' in backend:", exc_info=True)
                 return ret
 
             if request.method != "GET":
@@ -156,8 +160,8 @@ def cache(
 
             try:
                 await backend.set(cache_key, encoded_ret, expire)
-            except ConnectionError:
-                pass
+            except Exception:
+                logger.warning(f"Error setting cache key '{cache_key}' in backend:", exc_info=True)
 
             response.headers["Cache-Control"] = f"max-age={expire}"
             etag = f"W/{hash(encoded_ret)}"